refactor(#70): PostService 이미지관력 로직 분리
