+++ src/main/java/friendy/community/domain/post/service/PostService.java
@@ -10,13 +10,10 @@ import friendy.community.domain.post.dto.request.PostUpdateRequest;
 import friendy.community.domain.post.dto.response.FindAllPostResponse;
 import friendy.community.domain.post.dto.response.FindPostResponse;
 import friendy.community.domain.post.model.Post;
-import friendy.community.domain.post.model.PostImage;
-import friendy.community.domain.post.repository.PostImageRepository;
 import friendy.community.domain.post.repository.PostQueryDSLRepository;
 import friendy.community.domain.post.repository.PostRepository;
 import friendy.community.global.exception.ErrorCode;
 import friendy.community.global.exception.FriendyException;
-import friendy.community.infra.storage.s3.service.S3service;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
@@ -25,10 +22,7 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import java.util.HashSet;
 import java.util.List;
-import java.util.Optional;
-import java.util.Set;
 
 @Service
 @RequiredArgsConstructor
@@ -41,50 +35,33 @@ public class PostService {
     private final JwtTokenProvider jwtTokenProvider;
     private final AuthService authService;
     private final HashtagService hashtagService;
-    private final S3service s3service;
-    private final PostImageRepository postImageRepository;
+    private final PostImageService postImageService;
 
-    public long savePost(final PostCreateRequest postCreateRequest, final HttpServletRequest httpServletRequest) {
+    public long savePost(final PostCreateRequest request, final HttpServletRequest httpServletRequest) {
         final Member member = getMemberFromRequest(httpServletRequest);
-        final Post post = Post.of(postCreateRequest, member);
+        final Post post = Post.of(request, member);
 
-        if (postCreateRequest.imageUrls() != null) {
-            int imageOrder = 1;
-            for (String imageUrl : postCreateRequest.imageUrls()) {
-                PostImage postImage = savePostImage(imageUrl,imageOrder++);
-                postImage.assignPost(post);
-                post.addImage(postImage);
-            }
+        if (request.imageUrls() != null) {
+            postImageService.saveImagesForPost(post, request.imageUrls());
         }
 
         postRepository.save(post);
-        hashtagService.saveHashtags(post, postCreateRequest.hashtags());
+        hashtagService.saveHashtags(post, request.hashtags());
 
         return post.getId();
     }
 
-    public long updatePost(
-        final PostUpdateRequest postUpdateRequest,
-        final HttpServletRequest httpServletRequest,
-        final Long postId
-    ) {
+    public long updatePost(final PostUpdateRequest request, final HttpServletRequest httpServletRequest, final Long postId) {
         final Member member = getMemberFromRequest(httpServletRequest);
         final Post post = validatePostExistence(postId);
         validatePostAuthor(member, post);
 
-        if (postUpdateRequest.imageUrls() != null) {
-            List<PostImage> existingPostImages = postImageRepository.findByPostIdOrderByImageOrderAsc(postId);
-
-            Set<String> newImageUrlSet = new HashSet<>(postUpdateRequest.imageUrls());
-            deleteUnusedImages(existingPostImages, newImageUrlSet);
-
-            existingPostImages = postImageRepository.findByPostIdOrderByImageOrderAsc(postId);
-            updatePostImages(postUpdateRequest.imageUrls(), existingPostImages, post);
+        if (request.imageUrls() != null) {
+            postImageService.updateImagesForPost(post, request.imageUrls());
         }
-
-        post.updatePost(postUpdateRequest);
+        post.updatePost(request);
+        hashtagService.updateHashtags(post, request.hashtags());
         postRepository.save(post);
-        hashtagService.updateHashtags(post, postUpdateRequest.hashtags());
 
         return post.getId();
     }
@@ -94,11 +71,7 @@ public class PostService {
         final Post post = validatePostExistence(postId);
         validatePostAuthor(member, post);
 
-        List<PostImage> imagesToDelete = postImageRepository.findByPostIdOrderByImageOrderAsc(postId);
-        for(PostImage image : imagesToDelete) {
-            s3service.deleteFromS3(image.getS3Key());
-        }
-
+        postImageService.deleteImagesForPost(post);
         hashtagService.deleteHashtags(postId);
         postRepository.delete(post);
     }
@@ -143,51 +116,5 @@ public class PostService {
         final String accessToken = jwtTokenExtractor.extractAccessToken(httpServletRequest);
         final String email = jwtTokenProvider.extractEmailFromAccessToken(accessToken);
         return authService.getMemberByEmail(email);
-    }
-
-    private PostImage savePostImage(String requestImageUrl, int imageOrder) {
-        String imageUrl = s3service.moveS3Object(requestImageUrl, "post");
-        String s3Key = s3service.extractFilePath(imageUrl);
-        String fileType = s3service.getContentTypeFromS3(s3Key);
-        return PostImage.of(imageUrl, s3Key, fileType, imageOrder);
-    }
-
-    private void updatePostImages(List<String> newImageUrls, List<PostImage> existingPostImages, Post post) {
-        int imageOrder = 1;
-
-        for (String newImageUrl : newImageUrls) {
-            Optional<PostImage> existingImageOpt = findExistingImage(newImageUrl, existingPostImages);
-
-            if (existingImageOpt.isPresent()) {
-                PostImage existingImage = existingImageOpt.get();
-                existingImage.changeImageOrder(imageOrder++);
-                postImageRepository.save(existingImage);
-            } else {
-                PostImage postNewImage = savePostImage(newImageUrl, imageOrder++);
-                postNewImage.assignPost(post);
-                post.addImage(postNewImage);
-            }
-        }
-    }
-
-    private void deleteUnusedImages(List<PostImage> existingPostImages, Set<String> newImageUrlSet) {
-        List<PostImage> imagesToRemove = findImagesToRemove(existingPostImages, newImageUrlSet);
-
-        for (PostImage imageToRemove : imagesToRemove) {
-            s3service.deleteFromS3(imageToRemove.getS3Key());
-            postImageRepository.delete(imageToRemove);
-        }
-    }
-
-    private Optional<PostImage> findExistingImage(String newImageUrl, List<PostImage> existingPostImages) {
-        return existingPostImages.stream()
-            .filter(image -> image.getImageUrl().equals(newImageUrl))
-            .findFirst();
-    }
-
-    private List<PostImage> findImagesToRemove(List<PostImage> existingPostImages, Set<String> newImageUrlSet) {
-        return existingPostImages.stream()
-            .filter(existingImage -> !newImageUrlSet.contains(existingImage.getImageUrl()))
-            .toList();
     }
 }
